[USB] Clean up devices interface
