+++ libraries/Mouse/src/Mouse.cpp
@@ -20,7 +20,6 @@
 */
 
 #include "Mouse.h"
-#include "usbd_hid_composite.h"
 
 #if defined(USBCON)
